requirements.txt
asgiref==3.5.2
Django==4.1.2
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.1
Pillow==9.2.0
psycopg2==2.9.4
PyJWT==1.7.1
python-decouple==3.6
pytz==2022.4
sqlparse==0.4.3
